Check user role before some actions

+++ src/ForumBundle/Controller/PostController.php
@@ -9,6 +9,7 @@ use FOS\RestBundle\Controller\Annotations as Rest;
 use Nelmio\ApiDocBundle\Annotation\Model;
 use Swagger\Annotations as SWG;
 use Symfony\Component\HttpFoundation\Request;
+use UserBundle\Entity\User;
 
 class PostController extends RestController
 {
@@ -78,6 +79,8 @@ class PostController extends RestController
      */
     public function postPostAction(Request $request)
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         return $this->processForm(new ForumPost(), $request, PostType::class, 'get_forum_post');
     }
 
@@ -100,12 +103,21 @@ class PostController extends RestController
      */
     public function putPostAction($id, Request $request)
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         $em = $this->getDoctrine()->getManager();
 
         $post = $em->getRepository('ForumBundle:Post')->find($id);
 
         if (null === $post) {
             throw $this->createNotFoundException();
+        }
+
+        /** @var User $user */
+        $user = $this->getUser();
+
+        if (!$this->isGranted('ROLE_ADMIN') && $user !== $post->getAuthor()) {
+            throw $this->createAccessDeniedException('This post does not belong to you');
         }
 
         return $this->processForm($post, $request, PostType::class);
@@ -130,12 +142,21 @@ class PostController extends RestController
      */
     public function deletePostAction($id)
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         $em = $this->getDoctrine()->getManager();
 
         $post = $em->getRepository('ForumBundle:Post')->find($id);
 
         if (null === $post) {
             throw $this->createNotFoundException();
+        }
+
+        /** @var User $user */
+        $user = $this->getUser();
+
+        if (!$this->isGranted('ROLE_ADMIN') && $user !== $post->getAuthor()) {
+            throw $this->createAccessDeniedException('This post does not belong to you');
         }
 
         $em->remove($post);
